Started MPI homework
I wrote code to set up a topology, and set up random seeds on each proc. The random seeds are done by having the root generate an array of random numbers (based on the time). These are then scattered to each proc/serve as the seed for that proc. For the topologies, I used a grid topology. The topology has pre-defined dimensions for 2,4,8,16,32, and 64 procs. It will not work for other numbers of procs. I had each proc initialize its random array using the received seed. I then send the border elements between adjacent procs (in the topology). A proc only tries to send/recv its border if it has a neighbor in the grid in a particular direction. This sharing of borders is done so that the procs are smoothing one giant grid (each proc is just working on a block of a greater grid, the border correction is done so that the edges of each block line up).

This code has been written but not compiled/debuged. The code is not finished (still need to parallelize the smoothing, counting, and timing part), then run tests to make sure the border stitching is working.