Modified timing, printout
I reformatted the output. Changes include replacing tabs with spaces (for portability) and changing the formatting of floating point numbers (proportions were printed in sci notation, timings with fewer decimal places). This was done to make the output more readible.

I also moved the reduction operations into the count timings, since we don't really have an accruate count until the reduction is finished/a global count is on the root.