Improved timing measurements
Rather than using timing measurements on the root process, I had each process calculate timing data. This data is then reduced and averaged on the root process. I think that this gives me a more relilable measure of the execution time (rather than just using the execution time on the root).

I also updated the switch statement for dimensions. The error statement in the default case only prints on the root now.